Homework #5 Добавлены классы данных для Movies Добавлен лист Movies Добавлена передача данных из фрагмента списка в детальный фрагмент Добавлен механизм сохранения (на время работы приложения) лайков и рейтингов Добавлен лист для актеров в детальном фрагменте с ListAdapter
