+++ app/src/main/java/com/example/superpupermegaproject/adapters/ActorsListAdapter.kt
@@ -1,13 +1,16 @@
-package com.example.superpupermegaproject
+package com.example.superpupermegaproject.adapters
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.example.superpupermegaproject.R
+import com.example.superpupermegaproject.model.Actor
 
 class ActorsListAdapter :
     ListAdapter<Actor, ActorsListAdapter.ActorsViewHolder>(ActorsDiffUtilCallback()) {
@@ -38,5 +41,15 @@ class ActorsListAdapter :
 
             tvActorName.setText(actor.name)
         }
+    }
+
+    class ActorsDiffUtilCallback : DiffUtil.ItemCallback<Actor>() {
+        override fun areItemsTheSame(oldItem: Actor, newItem: Actor): Boolean =
+            oldItem.name == newItem.name
+
+
+        override fun areContentsTheSame(oldItem: Actor, newItem: Actor): Boolean =
+            oldItem == newItem
+
     }
 }
