+++ app/src/main/java/com/example/superpupermegaproject/adapters/MoviesListAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.superpupermegaproject
+package com.example.superpupermegaproject.adapters
 
 import android.view.LayoutInflater
 import android.view.View
@@ -8,6 +8,11 @@ import android.widget.RatingBar
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.example.superpupermegaproject.R
+import com.example.superpupermegaproject.extensions.setTint
+import com.example.superpupermegaproject.model.Movie
+import com.example.superpupermegaproject.model.MoviesDataSource
+import com.example.superpupermegaproject.ui.MoviesListFragment
 
 class MoviesListAdapter: RecyclerView.Adapter<MoviesListAdapter.MovieViewHolder>() {
 
@@ -20,11 +25,11 @@ class MoviesListAdapter: RecyclerView.Adapter<MoviesListAdapter.MovieViewHolder>
         notifyDataSetChanged()
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MoviesListAdapter.MovieViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MovieViewHolder {
         return MovieViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.view_holder_movie, parent, false))
     }
 
-    override fun onBindViewHolder(holder: MoviesListAdapter.MovieViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: MovieViewHolder, position: Int) {
         holder.bind(moviesList[position])
         onClickListener?.let {listener ->
             holder.itemView.setOnClickListener { listener.onClickItem(moviesList[position].id) }
