+++ app/src/main/java/com/example/superpupermegaproject/model/Actor.kt
@@ -1,4 +1,4 @@
-package com.example.superpupermegaproject
+package com.example.superpupermegaproject.model
 
 data class Actor(val name: String,
                 val imageID: Int)
