+++ app/src/main/java/com/example/superpupermegaproject/model/MoviesDataSource.kt
@@ -1,4 +1,6 @@
-package com.example.superpupermegaproject
+package com.example.superpupermegaproject.model
+
+import com.example.superpupermegaproject.R
 
 object MoviesDataSource {
     val moviesList = listOf(
