+++ app/src/main/java/com/example/superpupermegaproject/ui/MoviesListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.superpupermegaproject
+package com.example.superpupermegaproject.ui
 
 import android.content.Context
 import android.os.Bundle
@@ -8,6 +8,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.superpupermegaproject.model.MoviesDataSource
+import com.example.superpupermegaproject.adapters.MoviesListAdapter
+import com.example.superpupermegaproject.R
 
 
 class MoviesListFragment : Fragment() {
